Remove the client.start method, too
This means that the client starts up immediately on creation.  It will
connect and set up a timer to reconnect, thus causing Node not to exit
if not handled correctly.  This also sets up a bit of a race to call
`client.on('connection', ..)`, but that race really only exists in the
tests.